fix: 修复显示设置恢复的边界场景（启动失败/重叠启动/快照不完整）
- CaptureAndSetWindowed: 读不到完整快照时跳过，避免修改后无法恢复
- CaptureAndSetWindowed: 已有未恢复快照时不再覆盖（TryAdd + 锁）
- 恢复监听前移到写注册表之后立即启动（SystemControl），启动失败时也能恢复
- 移除 HomePageViewModel 中的恢复监听，逻辑内聚到 SystemControl
